requirements.txt
balethon
python-dotenv
telethon